# src/main/resources/openapi/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:26000
  description: Generated server url
paths:
  /persons/{id}:
    get:
      tags:
      - person-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
    put:
      tags:
      - person-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Person'
        required: true
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Person'
    delete:
      tags:
      - person-controller
      operationId: deleteById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
components:
  schemas:
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        street:
          type: string
        houseNumber:
          type: string
        postalCode:
          type: string
        city:
          type: string
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        addresses:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Address'